refactor migrations

+++ src/main/resources/db/migration/V5__create_table_post_tags.sql
@@ -1,6 +1,6 @@
 create TABLE IF NOT EXISTS post_tags(
-  post_id INT,
-  tag_id INT,
+  post_id           BIGINT,
+  tag_id            BIGINT,
   PRIMARY KEY (post_id, tag_id),
   CONSTRAINT fk_post FOREIGN KEY(post_id) REFERENCES post(id),
   CONSTRAINT fk_tag FOREIGN KEY(tag_id) REFERENCES tag(id)
